+++ tests/testthat/test-cacheGeo.R
@@ -136,7 +136,7 @@ test_that("lightweight tests for code coverage", {
 
   outSF <- sf::st_as_sf(out)
 
-  skip_if_service_account_releaseVer_NotLinux()
+  skip_if_not_releaseVer_Linux()
 
   gls <- googledrive::drive_ls(cloudFolderID)
   alreadyThere <- gls$name %in% c(targetFile, targetFile2)